Move listener to presenter layer.
