+++ src/main/java/oripa/gui/presenter/creasepattern/SelectAllLineActionListener.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.application.main;
+package oripa.gui.presenter.creasepattern;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
